src/main/java/ua/epam/spring/hometask/service/impl/EventServiceImpl.java
package ua.epam.spring.hometask.service.impl;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.beans.factory.annotation.Qualifier;
import org.springframework.stereotype.Service;
import ua.epam.spring.hometask.domain.Event;
import ua.epam.spring.hometask.dao.EventDao;
import ua.epam.spring.hometask.service.EventService;

import javax.annotation.Nonnull;
import javax.annotation.Nullable;
import java.util.Collection;

@Service("EventService")
public class EventServiceImpl implements EventService {

    private EventDao eventDao;


    @Autowired()
    @Qualifier("DerbyEventDao")
    public void setEventDao(EventDao eventDao) {
        this.eventDao = eventDao;
    }

    @Nullable
    @Override
    public Event getByName(@Nonnull String name) {
        return eventDao.getByName(name);
    }

    @Override
    public Event save(@Nonnull Event event) {
        return eventDao.saveEvent(event);
    }

    @Override
    public void remove(@Nonnull Event event) {
        eventDao.removeEvent(event);
    }

    @Override
    public Event getById(@Nonnull Long id) {
        return eventDao.getById(id);
    }

    @Nonnull
    @Override
    public Collection<Event> getAll() {
        return eventDao.getAll();
    }
}
